refactor: Organize grouping.rs into focused submodules
Split the 752-line grouping.rs file into a well-organized module structure:

- `grouping/aggregates.rs` (681 lines): AggregateAccumulator enum,
  accumulate/finalize logic, SQL value helpers, and comprehensive tests
- `grouping/hash.rs` (72 lines): Hash-based GROUP BY implementation
  with O(1) group lookups
- `grouping/mod.rs` (13 lines): Public API and module coordination

Benefits:
- Clear separation of concerns (aggregate logic vs. grouping strategies)
- Easier to maintain and extend (e.g., adding sorted grouping in the future)
- All existing imports remain compatible via re-exports
- All tests pass without modification

Closes #2279
